% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_update_date}
\alias{get_update_date}
\title{Get most recent update date of the COVID-19 Canada Open Data Working Group dataset}
\usage{
get_update_date()
}
\description{
Get most recent update date of the COVID-19 Canada Open Data Working Group dataset
}
