% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load_Covid19Canada}
\alias{load_Covid19Canada}
\title{Load data from the COVID-19 Canada Open Data Working Group dataset}
\usage{
load_Covid19Canada(x = c("all", "all_prov", "all_hr"))
}
\arguments{
\item{x}{Character vector of the datasets to load: "all" (all time series), "all_prov" (all provincial time series), "all_hr" (all health region time series), or a vector of individual file names.}
}
\description{
Load data from the COVID-19 Canada Open Data Working Group dataset
}
\details{
Valid datasets include:
\itemize{
\item cases_ts_prov
\item mortality_ts_prov
\item active_ts_prov
\item testing_ts_prov
\item cases_ts_hr
\item mortality_ts_hr
}
}
