% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends_rolling_average.R
\name{rolling_average}
\alias{rolling_average}
\title{Calculate rolling averages}
\usage{
rolling_average(x, window_days = 7)
}
\arguments{
\item{x}{The time series for which to calculate the rolling average (for column "val").}

\item{window_days}{The length of the window in days for the rolling average (default = 7).}
}
\value{
The time series with the rolling average added in new column "rolling_avg".
}
\description{
Calculate rolling averages
}
