% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends_rolling_average.R
\name{trends_rolling_average}
\alias{trends_rolling_average}
\title{Calculate and summarize percent change in rolling average in tweetable form}
\usage{
trends_rolling_average(
  x,
  stat = c("cases", "mortality", "active", "testing"),
  new_date = get_update_date(),
  loc = c("all_prov", "all_hr"),
  before_days = 7,
  window_days = 7,
  threshold = 10,
  print_abs = FALSE,
  file
)
}
\arguments{
\item{x}{The time series for which to calculate the rolling average.}

\item{stat}{The stat of the time series. One of "cases", "mortality", "active', or "testing".}

\item{new_date}{The date for which to calculate the percent change since the previous date (by default, the most recent date the dataset was updated).}

\item{loc}{The locations for which to calculate trends. One of "all_prov" (all provinces), "all_hr" (all health regions), a vector of 2-letter province codes, or a vector of 4-number health region codes.}

\item{before_days}{The number of days before \code{new_date} to compare with to calculate the percent change (default: 7).}

\item{window_days}{The length of the window in days for the rolling average (default = 7).}

\item{threshold}{The percent threshold (in absolute value) for delineating a positive/negative trend from no change (default = 10).}

\item{print_abs}{Logical. Print absolute change in value from before date to new date? (Default = FALSE)}

\item{file}{A character string for the path and name of the output file.}

\item{print_abs_digits}{If \code{print_abs} is TRUE, the number of digits to print (defaults to 1).}
}
\value{
A text file summarizing the trends in the selected value over a particular time range and set of locations. Includes emojis.
}
\description{
Calculate and summarize percent change in rolling average in tweetable form
}
