% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trends_rolling_average.R
\name{trends_rolling_average}
\alias{trends_rolling_average}
\title{Calculate and summarize percent change in rolling average in tweetable form}
\usage{
trends_rolling_average(
  x,
  stat = c("cases", "mortality", "active", "testing"),
  new_date = get_update_date(),
  loc = c("all_prov", "all_hr"),
  before_days = 7,
  window_days = 7,
  threshold = 10,
  change_digits = 1,
  print_val = FALSE,
  print_val_digits = 1,
  min_val_before = NULL,
  min_val_new = NULL,
  output_lines = NULL,
  file
)
}
\arguments{
\item{x}{The time series for which to calculate the rolling average.}

\item{stat}{The stat of the time series. One of "cases", "mortality", "active", or "testing".}

\item{new_date}{The date for which to calculate the percent change since the previous date (by default, the most recent date the dataset was updated).}

\item{loc}{The locations for which to calculate trends. One of "all_prov" (all provinces), "all_hr" (all health regions), a vector of 2-letter province codes, or a vector of 4-number health region codes.}

\item{before_days}{The number of days before \code{new_date} to compare with to calculate the percent change (default: 7).}

\item{window_days}{The length of the window in days for the rolling average (default = 7).}

\item{threshold}{The percent threshold (in absolute value) for delineating a positive/negative trend from no change (default = 10).}

\item{change_digits}{The number of digits to print for the percent change (defaults to 1).}

\item{print_val}{Logical. Also print absolute changes in value from before date to new date? (Default = FALSE)}

\item{print_val_digits}{The number of digits to print for absolute values (defaults to 1). Used if \code{print_val} is TRUE and when a location goes from 0/day on the before date to > 0/day on the new date.}

\item{min_val_before}{Optional. If set, remove locations with values below the stated number on the before date (see Details).}

\item{min_val_new}{Optional. If set, remove locations with values below the stated number on the new date (see Details).}

\item{output_lines}{Optional. The number of lines to output (default: keep all lines).}

\item{file}{A character string for the path and name of the output file.}
}
\value{
A text file summarizing the trends in the selected value over a particular time range and set of locations. Includes emojis.
}
\description{
Calculate and summarize percent change in rolling average in tweetable form
}
\details{
The arguments \code{min_val_before} and \code{min_val_new} can be used to censor small absolute changes that correspond to huge percent changes. For example, suppose a health region goes from 0.1 cases/day -> 0.6 cases/day represents a 500\% increase but a small absolute change. Setting either \code{min_val_before} to a value > 0.1 or \code{min_val_new} to a value > 0.6 would censor this health region. Setting \code{min_val_before} to 0 implicitly censors locations with 0 to > 0 and 0 to 0, whereas setting \code{min_val_after} to 0 implicitly censors locations with percent change = -100\% and 0 to 0.
}
